WIP renderer rework for path tracing

+++ crosscompile/windows/clean.sh
@@ -1,23 +1,27 @@
-cd "`dirname $0`/../.."
-PROJECT_DIR=`pwd`
-cd "`dirname $0`"
-set -e
+ping -c 1 WINDOWS_PC > /dev/null
+if [[ $? == 0 ]] ; then
 
-# Kill potentially running process on remote windows pc
-ssh WINDOWS_PC "START /wait taskkill /f /im tests.exe"
-ssh WINDOWS_PC "START /wait taskkill /f /im demo.exe"
+	cd "`dirname $0`/../.."
+	PROJECT_DIR=`pwd`
+	cd "`dirname $0`"
+	set -e
 
-# Delete build on remote windows pc
-ssh WINDOWS_PC "for /d %i in (\v4d_build\debug\*) do @rmdir /s /q \"%i\" > NUL"
-ssh WINDOWS_PC "del /q /s \v4d_build\debug\* > NUL"
-ssh WINDOWS_PC "for /d %i in (\v4d_build\release\*) do @rmdir /s /q \"%i\" > NUL"
-ssh WINDOWS_PC "del /q /s \v4d_build\release\* > NUL"
+	# Kill potentially running process on remote windows pc
+	ssh WINDOWS_PC "START /wait taskkill /f /im tests.exe"
+	ssh WINDOWS_PC "START /wait taskkill /f /im demo.exe"
 
-# Cleanup local build
-rm -rf build
-mkdir build
+	# Delete build on remote windows pc
+	ssh WINDOWS_PC "for /d %i in (\v4d_build\debug\*) do @rmdir /s /q \"%i\" > NUL"
+	ssh WINDOWS_PC "del /q /s \v4d_build\debug\* > NUL"
+	ssh WINDOWS_PC "for /d %i in (\v4d_build\release\*) do @rmdir /s /q \"%i\" > NUL"
+	ssh WINDOWS_PC "del /q /s \v4d_build\release\* > NUL"
 
-# Copy global DLLs to Remote Windows PC
-scp -rq dll/*.dll WINDOWS_PC:/v4d_build/debug/
-scp -rq dll/*.dll WINDOWS_PC:/v4d_build/release/
+	# Cleanup local build
+	rm -rf build
+	mkdir build
 
+	# Copy global DLLs to Remote Windows PC
+	scp -rq dll/*.dll WINDOWS_PC:/v4d_build/debug/
+	scp -rq dll/*.dll WINDOWS_PC:/v4d_build/release/
+
+fi
