+++ crosscompile/windows/send.sh
@@ -1,19 +1,24 @@
-cd `dirname $0`/../../build
+ping -c 1 WINDOWS_PC > /dev/null
+if [[ $? == 0 ]] ; then
 
-ssh WINDOWS_PC "START /wait taskkill /f /im demo.exe" &&\
-ssh WINDOWS_PC "START /wait taskkill /f /im tests.exe" &&\
-ssh WINDOWS_PC "del /q /s C:\\v4d_build\\$1\\*.exe"
+	cd `dirname $0`/../../build
 
-echo "Copying files to remote windows machine..."
+	ssh WINDOWS_PC "START /wait taskkill /f /im demo.exe" &&\
+	ssh WINDOWS_PC "START /wait taskkill /f /im tests.exe" &&\
+	ssh WINDOWS_PC "del /q /s C:\\v4d_build\\$1\\*.exe"
 
-cd $1/modules
-for d in *; do
-	ssh WINDOWS_PC "if not exist C:\\v4d_build\\$1\\modules\\$d mkdir C:\\v4d_build\\$1\\modules\\$d"
-done
-cd ..
+	echo "Copying files to remote windows machine..."
 
-find -type f -name "*.exe" -exec scp -rq {''} WINDOWS_PC:/v4d_build/$1/{''} \;
-find -type f -name "*.dll" -exec scp -rq {''} WINDOWS_PC:/v4d_build/$1/{''} \;
-find -type d -name "assets" -exec scp -rq {''} WINDOWS_PC:/v4d_build/$1/{''} \;
+	cd $1/modules
+	for d in *; do
+		ssh WINDOWS_PC "if not exist C:\\v4d_build\\$1\\modules\\$d mkdir C:\\v4d_build\\$1\\modules\\$d"
+	done
+	cd ..
 
-echo "Ready!"
+	find -type f -name "*.exe" -exec scp -rq {''} WINDOWS_PC:/v4d_build/$1/{''} \;
+	find -type f -name "*.dll" -exec scp -rq {''} WINDOWS_PC:/v4d_build/$1/{''} \;
+	find -type d -name "assets" -exec scp -rq {''} WINDOWS_PC:/v4d_build/$1/{''} \;
+
+	echo "Ready!"
+
+fi
